Consolidate database queries into queries.ts and remove separate files
